html/browsers/browsing-the-web/back-forward-cache/events.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/html/cross-origin-embedder-policy/credentialless/resources/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
for (const originType of ['SameOrigin', 'SameSite', 'CrossSite']) {
promise_test(async t => {
  const idA = token();

  window.open(
    executorPath + idA + '&events=visibilitychange,pagehide,pageshow,load',
    '_blank', 'noopener');
  await navigate(idA, eval(`origin${originType}`) + backPath);

  await assert_bfcached(idA);

  assert_array_equals(
    await evalOn(idA, `getRecordedEvents()`),
    [
      'window.load',
      'window.pageshow',
      'window.pagehide.persisted',
      'document.visibilitychange.hidden',
      'window.visibilitychange.hidden',
      'document.visibilitychange.visible',
      'window.visibilitychange.visible',
      'window.pageshow.persisted'
    ]);
}, `Events fired (${originType})`);
}
</script>
